requirements.txt
flask
python-dotenv
alpaca-py 